requirements.txt
polars>=1.0.0
# H-2: Pin polars-hash to exact version. polars-hash hashes the string
# representation of values. A Polars or polars-hash version upgrade that
# changes how values render as strings will silently change hashes, causing
# every row to appear "changed" — a mass-update cascade. Run hash regression
# tests before upgrading either package.
polars-hash==0.4.5
connectorx>=0.3.3
oracledb>=2.0.0
pyodbc>=5.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
